Define the supported Rust API boundary
