+++ .github/supply-chain-tools.env
@@ -1,6 +1,7 @@
 # Release tooling is consumed only at these exact versions or OCI digests.
 # Update the workflow references and the policy tests in the same pull request.
 CARGO_DENY_VERSION=0.20.2
+CARGO_SEMVER_CHECKS_VERSION=0.49.0
 DIESEL_CLI_VERSION=2.3.11
 SYFT_IMAGE=anchore/syft:v1.50.0@sha256:1288ea4c8b38767b4e620c1e312c8cb26b6e887a99b4f07ab6cd19fc6f225026
 TRIVY_IMAGE=aquasec/trivy:0.73.0@sha256:7cced7cae583819fc7806d4cbc0dbbc7cad18b99f7d3e235192e6da8c091045c
