+++ crates/hubuum-event-sink-valkey/Cargo.toml
@@ -5,6 +5,10 @@ edition = "2024"
 description = "Valkey/Redis Streams event sink transport for Hubuum."
 license = "MIT"
 repository = "https://github.com/hubuum/hubuum"
+publish = false
+
+[package.metadata.hubuum]
+rust-api = "workspace-internal"
 
 [dependencies]
 futures = "0.3"
