+++ crates/hubuum-events-core/Cargo.toml
@@ -5,6 +5,10 @@ edition = "2024"
 description = "Backend-agnostic event catalog and provenance types for Hubuum's unified event & audit stream."
 license = "MIT"
 repository = "https://github.com/hubuum/hubuum"
+publish = false
+
+[package.metadata.hubuum]
+rust-api = "workspace-internal"
 
 [dependencies]
 chrono = { version = "0.4", features = ["serde"] }
