+++ crates/hubuum-templates/Cargo.toml
@@ -5,6 +5,10 @@ edition = "2024"
 description = "Bounded MiniJinja template helpers for Hubuum."
 license = "MIT"
 repository = "https://github.com/hubuum/hubuum"
+publish = false
+
+[package.metadata.hubuum]
+rust-api = "workspace-internal"
 
 [dependencies]
 chrono = "0.4"
