+++ src/lib.rs
@@ -1,4 +1,19 @@
+//! Internal application implementation for Hubuum's binaries.
+//!
+//! This crate is not a supported third-party embedding interface. Hubuum's
+//! supported application contract is the versioned HTTP API described by the
+//! generated OpenAPI document. Rust API consumers should use the separately
+//! maintained HTTP client. See `docs/rust_api_boundary.md` for the workspace
+//! crate classification and promotion policy.
+
 #![allow(async_fn_in_trait)]
+
+mod administration;
+mod application;
+
+pub use administration::run_admin_from_environment;
+pub use application::run_server_from_environment;
+pub use errors::ApiError;
 
 pub mod api;
 pub mod auth;
@@ -31,3 +46,14 @@ pub mod tls;
 pub mod token_retention;
 pub mod traits;
 pub mod utilities;
+
+/// Generate the canonical pretty-printed OpenAPI document.
+///
+/// This is the workspace-internal entrypoint used by `hubuum-openapi`.
+#[must_use]
+pub fn generated_openapi_json() -> String {
+    use utoipa::OpenApi;
+
+    let openapi = api::openapi::ApiDoc::openapi();
+    serde_json::to_string_pretty(&openapi).expect("failed to serialize OpenAPI document")
+}
